Updated plugin to use filter:topic.create (#48)
* Updated plugin to use filter:topic.create

Initially this plugin used a client-side toggle that would wait
until after a topic was posted, and then make a socket call to
toggle the question status. Write API couldn't hook into this,
so a server-side solution using filter:topic.create is required.

Also, I fixed issues with the "Ask a Question" label not showing
up in the composer dropdown.

Lastly, settings are now saved in a local variable called
plugin._settings. Side effect of this change is that a reload is
now required after changing ACP settings. Prior to this, the
settings were retrieved every time the page was reloaded.

* use parseInt instead of doublebang and trinary if